Add js file for validation form.
